+++ server/common/model/src/main/java/com/alipay/sofa/registry/common/model/slot/SlotFunction.java
@@ -22,6 +22,6 @@ package com.alipay.sofa.registry.common.model.slot;
  * @version v 0.1 2020-11-02 15:35 yuzhi.lyz Exp $
  */
 public interface SlotFunction {
-    short maxSlots();
-    short slotOf(Object o);
+    int maxSlots();
+    int slotOf(Object o);
 }
